Split React helpers from component modules
