+++ frontend/src/components/panels/TerminalPanel.tsx
@@ -5,8 +5,8 @@ import type { WebglAddon } from '@xterm/addon-webgl';
 import type { WebLinksAddon } from '@xterm/addon-web-links';
 import type { SerializeAddon } from '@xterm/addon-serialize';
 import type { Unicode11Addon } from '@xterm/addon-unicode11';
-import { useSession } from '../../contexts/SessionContext';
-import { useTheme } from '../../contexts/ThemeContext';
+import { useSession } from '../../contexts/useSession';
+import { useTheme } from '../../contexts/useTheme';
 import { TerminalPanelProps } from '../../types/panelComponents';
 import { isHotkeyEnabledForEvent, useHotkeyStore } from '../../stores/hotkeyStore';
 import { renderLog, devLog } from '../../utils/console';
