+++ frontend/src/components/panels/diff/CombinedDiffView.tsx
@@ -8,19 +8,9 @@ import type { ExecutionDiff, GitDiffResult } from '../../../types/diff';
 import { RefreshCw } from 'lucide-react';
 import { panelApi } from '../../../services/panelApi';
 import { usePanelStore } from '../../../stores/panelStore';
+import { clearPendingViewCommit, takePendingViewCommit } from './pendingViewCommit';
 
 const HISTORY_LIMIT = 50;
-
-// Module-level pending commit hash — survives mount/unmount cycles.
-// When the diff panel is not active its CombinedDiffView is unmounted,
-// so a synchronous CustomEvent would be lost. SessionView writes here
-// before dispatching the event; CombinedDiffView reads it on mount.
-let pendingViewCommit: { sessionId: string; commitHash: string } | null = null;
-
-/** Called by SessionView before dispatching 'diff:view-commit'. */
-export function setPendingViewCommit(sessionId: string, commitHash: string) {
-  pendingViewCommit = { sessionId, commitHash };
-}
 
 const SIDEBAR_STORAGE_KEY = 'diff-panel-sidebar-width';
 const DEFAULT_SIDEBAR_WIDTH = 300;
@@ -320,11 +310,11 @@ const CombinedDiffView = memo(forwardRef<CombinedDiffViewHandle, CombinedDiffVie
   // fired while this component was unmounted (non-active panels are not rendered).
   useEffect(() => {
     // Consume any pending hash written before this component mounted
-    if (pendingViewCommit && pendingViewCommit.sessionId === sessionId) {
+    const pendingCommitHash = takePendingViewCommit(sessionId);
+    if (pendingCommitHash !== null) {
       combinedDiffRequestIdRef.current += 1;
-      setViewingCommitHash(pendingViewCommit.commitHash);
+      setViewingCommitHash(pendingCommitHash);
       setSelectedExecutions([]);
-      pendingViewCommit = null;
     }
 
     const handler = (event: Event) => {
@@ -334,7 +324,7 @@ const CombinedDiffView = memo(forwardRef<CombinedDiffViewHandle, CombinedDiffVie
       combinedDiffRequestIdRef.current += 1;
       setViewingCommitHash(commitHash);
       setSelectedExecutions([]);
-      pendingViewCommit = null; // consumed
+      clearPendingViewCommit();
     };
     window.addEventListener('diff:view-commit', handler);
     return () => window.removeEventListener('diff:view-commit', handler);
