+++ frontend/src/contexts/portalContainerContextValue.ts
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const PortalContainerContext = createContext<HTMLElement | null>(null);
