+++ frontend/src/contexts/sessionContextValue.ts
@@ -1,0 +1,32 @@
+import { createContext } from 'react';
+import type { LucideIcon } from 'lucide-react';
+import type { GitCommands, Session } from '../types/session';
+
+export interface SessionContextValue {
+  sessionId: string;
+  workingDirectory: string;
+  projectId: string;
+  projectName?: string;
+  session: Session;
+  gitBranchActions?: Array<{
+    id: string;
+    label: string;
+    icon: LucideIcon;
+    onClick: () => void;
+    disabled: boolean;
+    variant: 'default' | 'success' | 'danger';
+    description: string;
+    shortcut?: string;
+    disabledReason?: string;
+  }>;
+  isMerging?: boolean;
+  gitCommands?: GitCommands;
+  onOpenIDEWithCommand?: (command?: string) => void;
+  onConfigureIDE?: () => void;
+  onSetTracking?: () => void;
+  trackingBranch?: string | null;
+  configuredIDECommand?: string | null;
+  isRemoteMode?: boolean;
+}
+
+export const SessionContext = createContext<SessionContextValue | undefined>(undefined);
