+++ frontend/src/contexts/themeContextValue.ts
@@ -1,0 +1,13 @@
+import { createContext } from 'react';
+
+export type Theme = 'light' | 'light-rounded' | 'dark' | 'oled' | 'dusk' | 'dusk-oled' | 'forge' | 'ember' | 'aurora' | 'night-owl' | 'night-owl-oled' | 'terracotta';
+
+export interface ThemeContextType {
+  theme: Theme;
+  setTheme: (theme: Theme) => void;
+  // Read-only here: the Appearance settings toggle writes it through
+  // persistence.saveConfig, and it flows back via the config sync effect.
+  highContrast: boolean;
+}
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
