+++ frontend/src/contexts/usePortalContainer.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { PortalContainerContext } from './portalContainerContextValue';
+
+export function usePortalContainer(): HTMLElement | null {
+  return useContext(PortalContainerContext);
+}
