+++ frontend/src/hooks/usePaneLogo.ts
@@ -1,6 +1,6 @@
 import paneLogoDark from '../assets/pane-logo-dark.svg';
 import paneLogoLight from '../assets/pane-logo-light.svg';
-import { useTheme } from '../contexts/ThemeContext';
+import { useTheme } from '../contexts/useTheme';
 
 export function usePaneLogo(): string {
   const { theme } = useTheme();
